+++ mod-source-record-manager-server/src/main/java/org/folio/services/migration/impl/AuthorityMappingGeographicSubdivisionCustomMigration.java
@@ -1,0 +1,41 @@
+package org.folio.services.migration.impl;
+
+import static org.folio.Record.RecordType.MARC_AUTHORITY;
+
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
+import org.folio.services.MappingRuleService;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthorityMappingGeographicSubdivisionCustomMigration extends BaseMappingRulesMigration {
+
+  private static final JsonArray SUBFIELDS = JsonArray.of("v", "x", "y", "z");
+  private static final String TARGET_181 = "geographicSubdivision";
+  private static final String TARGET_481 = "sftGeographicSubdivision";
+  private static final String TARGET_581 = "saftGeographicSubdivision";
+  private static final String DESCRIPTION_181 = "Heading geographic subdivision";
+  private static final String DESCRIPTION_481 = "See from tracing geographic subdivision";
+  private static final String DESCRIPTION_581 = "See also from tracing geographic subdivision";
+  private static final String TAG_181 = "181";
+  private static final String TAG_481 = "481";
+  private static final String TAG_581 = "581";
+  private static final String FEATURE_VERSION = "3.10.0";
+  private static final String DESCRIPTION = "Authority mapping rules: add rules for geographic subdivision fields";
+
+  protected AuthorityMappingGeographicSubdivisionCustomMigration(MappingRuleService mappingRuleService) {
+    super(MARC_AUTHORITY, FEATURE_VERSION, DESCRIPTION, mappingRuleService);
+  }
+
+  @Override
+  protected String updateRules(JsonObject rules) {
+    var field181 = createField(TARGET_181, DESCRIPTION_181, SUBFIELDS, EMPTY_RULES);
+    var field481 = createField(TARGET_481, DESCRIPTION_481, SUBFIELDS, EMPTY_RULES);
+    var field581 = createField(TARGET_581, DESCRIPTION_581, SUBFIELDS, EMPTY_RULES);
+    addFieldIfNotExists(rules, TAG_181, field181);
+    addFieldIfNotExists(rules, TAG_481, field481);
+    addFieldIfNotExists(rules, TAG_581, field581);
+    return sortRules(rules).encode();
+  }
+}
+